+++ crates/moon/Cargo.toml
@@ -5,7 +5,7 @@ authors = ["Martin Kavík <martin@kavik.cz>"]
 edition = "2018"
 
 [dependencies]
-tokio = { version = "1", features = ["signal", "rt-multi-thread", "fs"], default-features = false }
+tokio = { version = "1", features = ["rt-multi-thread", "fs"], default-features = false }
 tokio-stream = { version = "0.1.3", default-features = false }
 futures = { version = "0.3.13", default-features = false }
 uuid = { version = "0.8", features = ["v4"], default-features = false }
